refactor(shared,asset-collection): get_logger / get_settings を lru_cache ベースに統一
引数注入パターン（logger / settings_instance を None デフォルトで受け取る）を廃止し、
lru_cache でキャッシュするシンプルなファクトリ関数に置き換える。
get_logger のインポートパスを src.config.settings から src.config に統一する。
