+++ lambda/asset-collection/src/config/__init__.py
@@ -1,4 +1,6 @@
+from shared.config.base_settings import get_logger
+
 from .asset_fetch_config import AssetFetchConfig
-from .settings import get_logger, get_settings
+from .settings import get_settings
 
 __all__ = ["AssetFetchConfig", "get_logger", "get_settings"]
